New option to exit with >0 code if old deps are found
Closes #39

----

BTW, in the README I found that some paragraphs were hard-wrapped and
some were not, so I wasn’t sure what to do. So I did the only logical
thing: introduced an additional wrapping style!

The wrapping style used in the new paragraphs is sometimes called
“semantic linebreaks” and sometimes “one sentence per line”

Some reading on this style, if you’re as geeky as me:

https://rhodesmill.org/brandon/2012/one-sentence-per-line/
https://nathangrigg.com/2012/05/strategic-line-breaks